add test_not_exits_album case to TestRemoveSongFromAlbum

+++ songs/services.py
@@ -30,10 +30,12 @@ def add_song_to_album(data: dict) -> models.Album:
 
 
 def remove_song_from_album(data: dict) -> models.Album:
-    album = get_album(id_=data["album"])
-    song = album.songs.filter(song__pk=data["song"])
-    if not song:
-        raise NotFound()
-    song.delete()
-    album.refresh_from_db()
-    return album
+    serializer = serializers.SongRemoveFromAlbumSerializer(data=data)
+    if serializer.is_valid(raise_exception=True):
+        album = get_album(id_=data["album"])
+        song = album.songs.filter(song__pk=data["song"])
+        if not song:
+            raise NotFound()
+        song.delete()
+        album.refresh_from_db()
+        return album
